+++ src/engine/box/box.tpp
@@ -1,7 +1,7 @@
 #include "engine/box/box.h"
 
 template <typename T> 
-void Box<T>::linkPos(std::pair<int, int>* pos){
+void Box<T>::linkPos(std::pair<double, double>* pos){
     this->pos = pos;
 }
